Use Empty Set Instead Of Null On Read Operations (#143)
We should make sets empty on read instead of null when there are no entries. This caught us out when we imported an organization with an empty members set and subsequently tried to set it to null. Terraform treats this as a type mismatch which prevented the job from completing.